Rearranging code: corrections and splitting tests after file renaming
