--- led-badge-12x48.py
@@ -1,1 +1,0 @@
-led-badge-11x44.py
